split table data into multible files
