+++ pyXSteam/tables/R7_97.py
@@ -1,0 +1,956 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Content of the Tables from related documents
+
+Sources:
+
+* IAPWS R7-97(2012) IAPWS Industrial formulation 1997 for the Thermodynamic
+Properties of Water and Steam, September 1997
+
+"""
+
+
+class R7_97:
+    """class with table data from IAPWS R7-97(2012)"""
+
+    # Table 1: Numerical values of the coefficients of the B23-equation, Eqs. (5) and (6), for defining the boundary between regions 2 and 3
+    Table1_n = [0.34805185628969e3, -0.11671859879975e1, 0.10192970039326e-2, 0.57254459862746e3, 0.13918839778870e2]
+
+    # Table 2: Numerical values of the coefficients and exponents of the dimensionless Gibbs free energy for region 1, Eq. (7)
+    Table2_I = [0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 4, 4, 4, 5, 8, 8, 21, 23, 29, 30, 31, 32]
+    Table2_J = [
+        -2,
+        -1,
+        0,
+        1,
+        2,
+        3,
+        4,
+        5,
+        -9,
+        -7,
+        -1,
+        0,
+        1,
+        3,
+        -3,
+        0,
+        1,
+        3,
+        17,
+        -4,
+        0,
+        6,
+        -5,
+        -2,
+        10,
+        -8,
+        -11,
+        -6,
+        -29,
+        -31,
+        -38,
+        -39,
+        -40,
+        -41,
+    ]
+    Table2_n = [
+        +0.14632971213167,
+        -0.84548187169114,
+        -0.37563603672040e1,
+        +0.33855169168385e1,
+        -0.95791963387872,
+        +0.15772038513228,
+        -0.16616417199501e-1,
+        +0.81214629983568e-3,
+        +0.28319080123804e-3,
+        -0.60706301565874e-3,
+        -0.18990068218419e-1,
+        -0.32529748770505e-1,
+        -0.21841717175414e-1,
+        -0.52838357969930e-4,
+        -0.47184321073267e-3,
+        -0.30001780793026e-3,
+        +0.47661393906987e-4,
+        -0.44141845330846e-5,
+        -0.72694996297594e-15,
+        -0.31679644845054e-4,
+        -0.28270797985312e-5,
+        -0.85205128120103e-9,
+        -0.22425281908000e-5,
+        -0.65171222895601e-6,
+        -0.14341729937924e-12,
+        -0.40516996860117e-6,
+        -0.12734301741641e-8,
+        -0.17424871230634e-9,
+        -0.68762131295531e-18,
+        +0.14478307828521e-19,
+        +0.26335781662795e-22,
+        -0.11947622640071e-22,
+        +0.18228094581404e-23,
+        -0.93537087292458e-25,
+    ]
+
+    # Table 6: Numerical values of the coefficients and exponents of the backward equation T ( p,h ) for region 1, Eq. (11)
+    Table6_I = [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 2, 2, 3, 3, 4, 5, 6]
+    Table6_J = [0, 1, 2, 6, 22, 32, 0, 1, 2, 3, 4, 10, 32, 10, 32, 10, 32, 32, 32, 32]
+    Table6_n = [
+        -0.23872489924521e3,
+        +0.40421188637945e3,
+        +0.11349746881718e3,
+        -0.58457616048039e1,
+        -0.15285482413140e-3,
+        -0.10866707695377e-5,
+        -0.13391744872602e2,
+        +0.43211039183559e2,
+        -0.54010067170506e2,
+        +0.30535892203916e2,
+        -0.65964749423638e1,
+        +0.93965400878363e-2,
+        +0.11573647505340e-6,
+        -0.25858641282073e-4,
+        -0.40644363084799e-8,
+        +0.66456186191635e-7,
+        +0.80670734103027e-10,
+        -0.93477771213947e-12,
+        +0.58265442020601e-14,
+        -0.15020185953503e-16,
+    ]
+
+    # Table 8: Numerical values of the coefficients and exponents of the backward equation T ( p,s ) for region 1, Eq. (13)
+    Table8_I = [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 4]
+    Table8_J = [0, 1, 2, 3, 11, 31, 0, 1, 2, 3, 12, 31, 0, 1, 2, 9, 31, 10, 32, 32]
+    Table8_n = [
+        +0.17478268058307e3,
+        +0.34806930892873e2,
+        +0.65292584978455e1,
+        +0.33039981775489,
+        -0.19281382923196e-6,
+        -0.24909197244573e-22,
+        -0.26107636489332,
+        +0.22592965981586,
+        -0.64256463395226e-1,
+        +0.78876289270526e-2,
+        +0.35672110607366e-9,
+        +0.17332496994895e-23,
+        +0.56608900654837e-3,
+        -0.32635483139717e-3,
+        +0.44778286690632e-4,
+        -0.51322156908507e-9,
+        -0.42522657042207e-25,
+        +0.26400441360689e-12,
+        +0.78124600459723e-28,
+        -0.30732199903668e-30,
+    ]
+
+    # Region2
+    # Table 10. Numerical values of the coefficients and exponents of the ideal-gas part
+    # gamma o of the dimensionless Gibbs free energy for region 2, Eq. (16)a
+    Table10_J0 = [0, 1, -5, -4, -3, -2, -1, 2, 3]
+    Table10_n0_a = [
+        -0.96927686500217e1,  # 1a
+        +0.10086655968018e2,  # 2a
+        -0.56087911283020e-2,
+        +0.71452738081455e-1,
+        -0.40710498223928,
+        +0.14240819171444e1,
+        -0.43839511319450e1,
+        -0.28408632460772,
+        +0.21268463753307e-1,
+    ]
+    # Numerical values of the coefficients and exponents of the ideal-gas part gamma o of
+    # the dimensionless Gibbs free energy for metastable region 2, Eq. (18)
+    Table10_n0_b = [
+        -0.96937268393049e1,  # 1b
+        +0.10087275970006e2,  # 2b
+        -0.56087911283020e-2,
+        +0.71452738081455e-1,
+        -0.40710498223928,
+        +0.14240819171444e1,
+        -0.43839511319450e1,
+        -0.28408632460772,
+        +0.21268463753307e-1,
+    ]
+
+    # Table 11: Numerical values of the coefficients and exponents of the residual part gamma r of
+    # the dimensionless Gibbs free energy for region 2, Eq. (17)
+    Table11_I = [
+        1,
+        1,
+        1,
+        1,
+        1,
+        2,
+        2,
+        2,
+        2,
+        2,
+        3,
+        3,
+        3,
+        3,
+        3,
+        4,
+        4,
+        4,
+        5,
+        6,
+        6,
+        6,
+        7,
+        7,
+        7,
+        8,
+        8,
+        9,
+        10,
+        10,
+        10,
+        16,
+        16,
+        18,
+        20,
+        20,
+        20,
+        21,
+        22,
+        23,
+        24,
+        24,
+        24,
+    ]
+    Table11_J = [
+        0,
+        1,
+        2,
+        3,
+        6,
+        1,
+        2,
+        4,
+        7,
+        36,
+        0,
+        1,
+        3,
+        6,
+        35,
+        1,
+        2,
+        3,
+        7,
+        3,
+        16,
+        35,
+        0,
+        11,
+        25,
+        8,
+        36,
+        13,
+        4,
+        10,
+        14,
+        29,
+        50,
+        57,
+        20,
+        35,
+        48,
+        21,
+        53,
+        39,
+        26,
+        40,
+        58,
+    ]
+    Table11_n = [
+        -0.17731742473213e-2,
+        -0.17834862292358e-1,
+        -0.45996013696365e-1,
+        -0.57581259083432e-1,
+        -0.50325278727930e-1,
+        -0.33032641670203e-4,
+        -0.18948987516315e-3,
+        -0.39392777243355e-2,
+        -0.43797295650573e-1,
+        -0.26674547914087e-4,
+        +0.20481737692309e-7,
+        +0.43870667284435e-6,
+        -0.32277677238570e-4,
+        -0.15033924542148e-2,
+        -0.40668253562649e-1,
+        -0.78847309559367e-9,
+        +0.12790717852285e-7,
+        +0.48225372718507e-6,
+        +0.22922076337661e-5,
+        -0.16714766451061e-10,
+        -0.21171472321355e-2,
+        -0.23895741934104e2,
+        -0.59059564324270e-17,
+        -0.12621808899101e-5,
+        -0.38946842435739e-1,
+        +0.11256211360459e-10,
+        -0.82311340897998e1,
+        +0.19809712802088e-7,
+        +0.10406965210174e-18,
+        -0.10234747095929e-12,
+        -0.10018179379511e-8,
+        -0.80882908646985e-10,
+        +0.10693031879409,
+        -0.33662250574171,
+        +0.89185845355421e-24,
+        +0.30629316876232e-12,
+        -0.42002467698208e-5,
+        -0.59056029685639e-25,
+        +0.37826947613457e-5,
+        -0.12768608934681e-14,
+        +0.73087610595061e-28,
+        +0.55414715350778e-16,
+        -0.94369707241210e-6,
+    ]
+
+    # Table 16: Numerical values of the coefficients and exponents of the residual part gamma r of
+    # the dimensionless Gibbs free energy for the metastable-vapor region, Eq. (19)
+    Table16_I = [1, 1, 1, 1, 2, 2, 2, 3, 3, 4, 4, 5, 5]
+    Table16_J = [0, 2, 5, 11, 1, 7, 16, 4, 16, 7, 10, 9, 10]
+    Table16_n = [
+        -0.73362260186506e-2,
+        -0.88223831943146e-1,
+        -0.72334555213245e-1,
+        -0.40813178534455e-2,
+        +0.20097803380207e-2,
+        -0.53045921898642e-1,
+        -0.76190409086970e-2,
+        -0.63498037657313e-2,
+        -0.86043093028588e-1,
+        +0.75321581522770e-2,
+        -0.79238375446139e-2,
+        -0.22888160778447e-3,
+        -0.26456501482810e-2,
+    ]
+
+    # Table 19: Numerical values of the coefficients of the B2bc-equation, Eqs. (20) and (21), for
+    # defining the boundary between subregions 2b and 2c with respect to T( p,h ) calculations
+    Table19_n = [
+        +0.90584278514723e3,
+        -0.67955786399241,
+        +0.12809002730136e-3,
+        +0.26526571908428e4,
+        +0.45257578905948e1,
+    ]
+
+    # Table 20: Numerical values of the coefficients and exponents of the backward
+    # equation T ( p,h ) for subregion 2a, Eq. (22)
+    Table20_I = [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 5, 6, 6, 7]
+    Table20_J = [0, 1, 2, 3, 7, 20, 0, 1, 2, 3, 7, 9, 11, 18, 44, 0, 2, 7, 36, 38, 40, 42, 44, 24, 44, 12, 32, 44, 32, 36, 42, 34, 44, 28]
+    Table20_n = [
+        +0.10898952318288e4,
+        +0.84951654495535e3,
+        -0.10781748091826e3,
+        +0.33153654801263e2,
+        -0.74232016790248e1,
+        +0.11765048724356e2,
+        +0.18445749355790e1,
+        -0.41792700549624e1,
+        +0.62478196935812e1,
+        -0.17344563108114e2,
+        -0.20058176862096e3,
+        +0.27196065473796e3,
+        -0.45511318285818e3,
+        +0.30919688604755e4,
+        +0.25226640357872e6,
+        -0.61707422868339e-2,
+        -0.31078046629583,
+        +0.11670873077107e2,
+        +0.12812798404046e9,
+        -0.98554909623276e9,
+        +0.28224546973002e10,
+        -0.35948971410703e10,
+        +0.17227349913197e10,
+        -0.13551334240775e5,
+        +0.12848734664650e8,
+        +0.13865724283226e1,
+        +0.23598832556514e6,
+        -0.13105236545054e8,
+        +0.73999835474766e4,
+        -0.55196697030060e6,
+        +0.37154085996233e7,
+        +0.19127729239660e5,
+        -0.41535164835634e6,
+        -0.62459855192507e2,
+    ]
+
+    # Table 21: Numerical values of the coefficients and exponents of the backward equation T ( p,h ) for subregion 2b, Eq. (23)
+    Table21_I = [0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 4, 4, 4, 5, 5, 5, 6, 7, 7, 9, 9]
+    Table21_J = [
+        0,
+        1,
+        2,
+        12,
+        18,
+        24,
+        28,
+        40,
+        0,
+        2,
+        6,
+        12,
+        18,
+        24,
+        28,
+        40,
+        2,
+        8,
+        18,
+        40,
+        1,
+        2,
+        12,
+        24,
+        2,
+        12,
+        18,
+        24,
+        28,
+        40,
+        18,
+        24,
+        40,
+        28,
+        2,
+        28,
+        1,
+        40,
+    ]
+    Table21_n = [
+        +0.14895041079516e4,
+        +0.74307798314034e3,
+        -0.97708318797837e2,
+        +0.24742464705674e1,
+        -0.63281320016026,
+        +0.11385952129658e1,
+        -0.47811863648625,
+        +0.85208123431544e-2,
+        +0.93747147377932,
+        +0.33593118604916e1,
+        +0.33809355601454e1,
+        +0.16844539671904,
+        +0.73875745236695,
+        -0.47128737436186,
+        +0.15020273139707,
+        -0.21764114219750e-2,
+        -0.21810755324761e-1,
+        -0.10829784403677,
+        -0.46333324635812e-1,
+        +0.71280351959551e-4,
+        +0.11032831789999e-3,
+        +0.18955248387902e-3,
+        +0.30891541160537e-2,
+        +0.13555504554949e-2,
+        +0.28640237477456e-6,
+        -0.10779857357512e-4,
+        -0.76462712454814e-4,
+        +0.14052392818316e-4,
+        -0.31083814331434e-4,
+        -0.10302738212103e-5,
+        +0.28217281635040e-6,
+        +0.12704902271945e-5,
+        +0.73803353468292e-7,
+        -0.11030139238909e-7,
+        -0.81456365207833e-13,
+        -0.25180545682962e-10,
+        -0.17565233969407e-17,
+        +0.86934156344163e-14,
+    ]
+
+    # Table 22: Numerical values of the coefficients and exponents of the backward equation T ( p,h ) for subregion 2c, Eq. (24)
+    Table22_I = [-7, -7, -6, -6, -5, -5, -2, -2, -1, -1, 0, 0, 1, 1, 2, 6, 6, 6, 6, 6, 6, 6, 6]
+    Table22_J = [0, 4, 0, 2, 0, 2, 0, 1, 0, 2, 0, 1, 4, 8, 4, 0, 1, 4, 10, 12, 16, 20, 22]
+    Table22_n = [
+        -0.32368398555242e13,
+        +0.73263350902181e13,
+        +0.35825089945447e12,
+        -0.58340131851590e12,
+        -0.10783068217470e11,
+        +0.20825544563171e11,
+        +0.61074783564516e6,
+        +0.85977722535580e6,
+        -0.25745723604170e5,
+        +0.31081088422714e5,
+        +0.12082315865936e4,
+        +0.48219755109255e3,
+        +0.37966001272486e1,
+        -0.10842984880077e2,
+        -0.45364172676660e-1,
+        +0.14559115658698e-12,
+        +0.11261597407230e-11,
+        -0.17804982240686e-10,
+        +0.12324579690832e-6,
+        -0.11606921130984e-5,
+        +0.27846367088554e-4,
+        -0.59270038474176e-3,
+        +0.12918582991878e-2,
+    ]
+
+    # Table 25: Numerical values of the coefficients and exponents of the backward equation T( p,s ) for subregion 2a, Eq. (25)
+    Table25_I = [
+        -1.5,
+        -1.5,
+        -1.5,
+        -1.5,
+        -1.5,
+        -1.5,
+        -1.25,
+        -1.25,
+        -1.25,
+        -1,
+        -1,
+        -1,
+        -1,
+        -1,
+        -1,
+        -0.75,
+        -0.75,
+        -0.5,
+        -0.5,
+        -0.5,
+        -0.5,
+        -0.25,
+        -0.25,
+        -0.25,
+        -0.25,
+        0.25,
+        0.25,
+        0.25,
+        0.25,
+        0.5,
+        0.5,
+        0.5,
+        0.5,
+        0.5,
+        0.5,
+        0.5,
+        0.75,
+        0.75,
+        0.75,
+        0.75,
+        1,
+        1,
+        1.25,
+        1.25,
+        1.5,
+        1.5,
+    ]
+    Table25_J = [
+        -24,
+        -23,
+        -19,
+        -13,
+        -11,
+        -10,
+        -19,
+        -15,
+        -6,
+        -26,
+        -21,
+        -17,
+        -16,
+        -9,
+        -8,
+        -15,
+        -14,
+        -26,
+        -13,
+        -9,
+        -7,
+        -27,
+        -25,
+        -11,
+        -6,
+        1,
+        4,
+        8,
+        11,
+        0,
+        1,
+        5,
+        6,
+        10,
+        14,
+        16,
+        0,
+        4,
+        9,
+        17,
+        7,
+        18,
+        3,
+        15,
+        5,
+        18,
+    ]
+    Table25_n = [
+        -0.39235983861984e6,
+        +0.51526573827270e6,
+        +0.40482443161048e5,
+        -0.32193790923902e3,
+        +0.96961424218694e2,
+        -0.22867846371773e2,
+        -0.44942914124357e6,
+        -0.50118336020166e4,
+        +0.35684463560015,
+        +0.44235335848190e5,
+        -0.13673388811708e5,
+        +0.42163260207864e6,
+        +0.22516925837475e5,
+        +0.47442144865646e3,
+        -0.14931130797647e3,
+        -0.19781126320452e6,
+        -0.23554399470760e5,
+        -0.19070616302076e5,
+        +0.55375669883164e5,
+        +0.38293691437363e4,
+        -0.60391860580567e3,
+        +0.19363102620331e4,
+        +0.42660643698610e4,
+        -0.59780638872718e4,
+        -0.70401463926862e3,
+        +0.33836784107553e3,
+        +0.20862786635187e2,
+        +0.33834172656196e-1,
+        -0.43124428414893e-4,
+        +0.16653791356412e3,
+        -0.13986292055898e3,
+        -0.78849547999872,
+        +0.72132411753872e-1,
+        -0.59754839398283e-2,
+        -0.12141358953904e-4,
+        +0.23227096733871e-6,
+        -0.10538463566194e2,
+        +0.20718925496502e1,
+        -0.72193155260427e-1,
+        +0.20749887081120e-6,
+        -0.18340657911379e-1,
+        +0.29036272348696e-6,
+        +0.21037527893619,
+        +0.25681239729999e-3,
+        -0.12799002933781e-1,
+        -0.82198102652018e-5,
+    ]
+
+    # Table 26: Numerical values of the coefficients and exponents of the backward equation T( p,s ) for subregion 2b, Eq. (26)
+    Table26_I = [
+        -6,
+        -6,
+        -5,
+        -5,
+        -4,
+        -4,
+        -4,
+        -3,
+        -3,
+        -3,
+        -3,
+        -2,
+        -2,
+        -2,
+        -2,
+        -1,
+        -1,
+        -1,
+        -1,
+        -1,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        2,
+        2,
+        2,
+        3,
+        3,
+        3,
+        4,
+        4,
+        5,
+        5,
+        5,
+    ]
+    Table26_J = [
+        0,
+        11,
+        0,
+        11,
+        0,
+        1,
+        11,
+        0,
+        1,
+        11,
+        12,
+        0,
+        1,
+        6,
+        10,
+        0,
+        1,
+        5,
+        8,
+        9,
+        0,
+        1,
+        2,
+        4,
+        5,
+        6,
+        9,
+        0,
+        1,
+        2,
+        3,
+        7,
+        8,
+        0,
+        1,
+        5,
+        0,
+        1,
+        3,
+        0,
+        1,
+        0,
+        1,
+        2,
+    ]
+    Table26_n = [
+        +0.31687665083497e6,
+        +0.20864175881858e2,
+        -0.39859399803599e6,
+        -0.21816058518877e2,
+        +0.22369785194242e6,
+        -0.27841703445817e4,
+        +0.99207436071480e1,
+        -0.75197512299157e5,
+        +0.29708605951158e4,
+        -0.34406878548526e1,
+        +0.38815564249115,
+        +0.17511295085750e5,
+        -0.14237112854449e4,
+        +0.10943803364167e1,
+        +0.89971619308495,
+        -0.33759740098958e4,
+        +0.47162885818355e3,
+        -0.19188241993679e1,
+        +0.41078580492196,
+        -0.33465378172097,
+        +0.13870034777505e4,
+        -0.40663326195838e3,
+        +0.41727347159610e2,
+        +0.21932549434532e1,
+        -0.10320050009077e1,
+        +0.35882943516703,
+        +0.52511453726066e-2,
+        +0.12838916450705e2,
+        -0.28642437219381e1,
+        +0.56912683664855,
+        -0.99962954584931e-1,
+        -0.32632037778459e-2,
+        +0.23320922576723e-3,
+        -0.15334809857450,
+        +0.29072288239902e-1,
+        +0.37534702741167e-3,
+        +0.17296691702411e-2,
+        -0.38556050844504e-3,
+        -0.35017712292608e-4,
+        -0.14566393631492e-4,
+        +0.56420857267269e-5,
+        +0.41286150074605e-7,
+        -0.20684671118824e-7,
+        +0.16409393674725e-8,
+    ]
+
+    # Table 27: Numerical values of the coefficient s and exponents of the backward equation T( p,s ) for subregion 2c, Eq. (27)
+    Table27_I = [-2, -2, -1, 0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 3, 3, 3, 4, 4, 4, 5, 5, 5, 6, 6, 7, 7, 7, 7, 7]
+    Table27_J = [0, 1, 0, 0, 1, 2, 3, 0, 1, 3, 4, 0, 1, 2, 0, 1, 5, 0, 1, 4, 0, 1, 2, 0, 1, 0, 1, 3, 4, 5]
+    Table27_n = [
+        +0.90968501005365e3,
+        +0.24045667088420e4,
+        -0.59162326387130e3,
+        +0.54145404128074e3,
+        -0.27098308411192e3,
+        +0.97976525097926e3,
+        -0.46966772959435e3,
+        +0.14399274604723e2,
+        -0.19104204230429e2,
+        +0.53299167111971e1,
+        -0.21252975375934e2,
+        -0.31147334413760,
+        +0.60334840894623,
+        -0.42764839702509e-1,
+        +0.58185597255259e-2,
+        -0.14597008284753e-1,
+        +0.56631175631027e-2,
+        -0.76155864584577e-4,
+        +0.22440342919332e-3,
+        -0.12561095013413e-4,
+        +0.63323132660934e-6,
+        -0.20541989675375e-5,
+        +0.36405370390082e-7,
+        -0.29759897789215e-8,
+        +0.10136618529763e-7,
+        +0.59925719692351e-11,
+        -0.20677870105164e-10,
+        -0.20874278181886e-10,
+        +0.10162166825089e-9,
+        -0.16429828281347e-9,
+    ]
+
+    # Region3
+
+    # Table 30: Numerical values of the coefficients and exponents of the dimensionless Helmholtz free energy for region 3, Eq. (28)
+    # TODO: check calculations for skip on index 0
+    Table30_I = [0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 4, 4, 4, 4, 5, 5, 5, 6, 6, 6, 7, 8, 9, 9, 10, 10, 11]
+    Table30_J = [
+        0,
+        0,
+        1,
+        2,
+        7,
+        10,
+        12,
+        23,
+        2,
+        6,
+        15,
+        17,
+        0,
+        2,
+        6,
+        7,
+        22,
+        26,
+        0,
+        2,
+        4,
+        16,
+        26,
+        0,
+        2,
+        4,
+        26,
+        1,
+        3,
+        26,
+        0,
+        2,
+        26,
+        2,
+        26,
+        2,
+        26,
+        0,
+        1,
+        26,
+    ]
+    Table30_n = [
+        +0.10658070028513e1,
+        -0.15732845290239e2,
+        +0.20944396974307e2,
+        -0.76867707878716e1,
+        +0.26185947787954e1,
+        -0.28080781148620e1,
+        +0.12053369696517e1,
+        -0.84566812812502e-2,
+        -0.12654315477714e1,
+        -0.11524407806681e1,
+        +0.88521043984318,
+        -0.64207765181607,
+        +0.38493460186671,
+        -0.85214708824206,
+        +0.48972281541877e1,
+        -0.30502617256965e1,
+        +0.39420536879154e-1,
+        +0.12558408424308,
+        -0.27999329698710,
+        +0.13899799569460e1,
+        -0.20189915023570e1,
+        -0.82147637173963e-2,
+        -0.47596035734923,
+        +0.43984074473500e-1,
+        -0.44476435428739,
+        +0.90572070719733,
+        +0.70522450087967,
+        +0.10770512626332,
+        -0.32913623258954,
+        -0.50871062041158,
+        -0.22175400873096e-1,
+        +0.94260751665092e-1,
+        +0.16436278447961,
+        -0.13503372241348e-1,
+        -0.14834345352472e-1,
+        +0.57922953628084e-3,
+        +0.32308904703711e-2,
+        +0.80964802996215e-4,
+        -0.16557679795037e-3,
+        -0.44923899061815e-4,
+    ]
+
+    # Region4
+
+    # Table 34: Numerical values of the coefficients of the dimensionless saturation equations, Eqs. (29) to (31)
+    Table34_n = [
+        +0.11670521452767e4,
+        -0.72421316703206e6,
+        -0.17073846940092e2,
+        +0.12020824702470e5,
+        -0.32325550322333e7,
+        +0.14915108613530e2,
+        -0.48232657361591e4,
+        +0.40511340542057e6,
+        -0.23855557567849,
+        +0.65017534844798e3,
+    ]
+
+    # Region5
+
+    # Table 37: Numerical values of the coefficients and exponents of the ideal-gas part gamma o of
+    # the dimensionless Gibbs free energy for region 5, Eq. (33)
+    Table37_J = [0, 1, -3, -2, -1, 2]
+    Table37_J = [0, 1, -3, -2, -1, 2]
+    Table37_n = [-0.13179983674201e2, +0.68540841634434e1, -0.24805148933466e-1, +0.36901534980333, -0.31161318213925e1, -0.329616265389177]
+
+    # Table 38: Numerical values of the coefficients and exponents of the residual part gamma r of
+    # the dimensionless Gibbs free energy for region 5, Eq. (34)
+    Table38_I = [1, 1, 1, 2, 2, 3]
+    Table38_J = [1, 2, 3, 3, 9, 8]
+    Table38_n = [
+        +0.15736404855259e-2,
+        +0.90153761673944e-3,
+        -0.50270077677648e-3,
+        +0.22440037409485e-5,
+        -0.41163275453471e-5,
+        +0.37919454822955e-7,
+    ]
+
+    # ?????????
+    # TODO find source
+    Tablexx_I = [1, 1, 1, 2, 3]
+    Tablexx_J = [0, 1, 3, 9, 3]
+    Tablexx_n = [
+        -1.2563183589592e-04,
+        2.1774678714571e-03,
+        -0.004594282089991,
+        -3.9724828359569e-06,
+        1.2919228289784e-07,
+    ]
