+++ pyXSteam/tables/SR2_01.py
@@ -1,0 +1,158 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Content of the Tables from related documents
+
+Sources:
+
+* IAPWS SR2-01(2014) Revised Supplementary Release on Backward Equations for
+Pressure as a Function of Enthalpy and Entropy p(h,s) for Regions 1 and 2 of 
+the IAPWS Industrial Formulation 1997 for the Thermodynamic Properties of 
+Water and Steam
+
+"""
+
+
+class SR2_01:
+    """class with table data from IAPWS SR2-01(2014)"""
+
+    # Table 2: Coefficients and exponents of the backward equation p_1(h,s) for region 1 in its dimensionless form, Eq. (1)
+    Table2_I = [0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 3, 4, 4, 5]
+    Table2_J = [0, 1, 2, 4, 5, 6, 8, 14, 0, 1, 4, 6, 0, 1, 10, 4, 1, 4, 0]
+    Table2_n = [
+        -0.691997014660582,
+        -0.183612548787560e2,
+        -0.928332409297335e1,
+        +0.659639569909906e2,
+        -0.162060388912024e2,
+        +0.450620017338667e3,
+        +0.854680678224170e3,
+        +0.607523214001162e4,
+        +0.326487682621856e2,
+        -0.269408844582931e2,
+        -0.319947848334300e3,
+        -0.928354307043320e3,
+        +0.303634537455249e2,
+        -0.650540422444146e2,
+        -0.430991316516130e4,
+        -0.747512324096068e3,
+        +0.730000345529245e3,
+        +0.114284032569021e4,
+        -0.436407041874559e3,
+    ]
+
+    # Table 5: Numerical values of the coefficients of the equation h2ab(s) in its dimensionless
+    # form, Eq. (2), for defining the boundary between subregions 2a and 2b
+    Table5_n = [-0.349898083432139e4, +0.257560716905876e4, -0.421073558227969e3, +0.276349063799944e2]
+
+    # Table 6: Coefficients and exponents of the backward equation p_2a(h,s) for subregion 2a
+    # in its dimensionless form, Eq. (3)
+    Table6_I = [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 3, 3, 3, 3, 3, 4, 5, 5, 6, 7]
+    Table6_J = [1, 3, 6, 16, 20, 22, 0, 1, 2, 3, 5, 6, 10, 16, 20, 22, 3, 16, 20, 0, 2, 3, 6, 16, 16, 3, 16, 3, 1]
+    Table6_n = [
+        -0.182575361923032e-1,
+        -0.125229548799536,
+        +0.592290437320145,
+        +0.604769706185122e1,
+        +0.238624965444474e3,
+        -0.298639090222922e3,
+        +0.512250813040750e-1,
+        -0.437266515606486,
+        +0.413336902999504,
+        -0.516468254574773e1,
+        -0.557014838445711e1,
+        +0.128555037824478e2,
+        +0.114144108953290e2,
+        -0.119504225652714e3,
+        -0.284777985961560e4,
+        +0.431757846408006e4,
+        +0.112894040802650e1,
+        +0.197409186206319e4,
+        +0.151612444706087e4,
+        +0.141324451421235e-1,
+        +0.585501282219601,
+        -0.297258075863012e1,
+        +0.594567314847319e1,
+        -0.623656565798905e4,
+        +0.965986235133332e4,
+        +0.681500934948134e1,
+        -0.633207286824489e4,
+        -0.558919224465760e1,
+        +0.400645798472063e-1,
+    ]
+
+    # Table 7: Coefficients and exponents of the backward equation p_2b(h,s) for subregion 2b in its dimensionless form, Eq. (4)
+    Table7_I = [0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 2, 2, 2, 3, 3, 3, 3, 4, 4, 5, 5, 6, 6, 6, 7, 7, 8, 8, 8, 8, 12, 14]
+    Table7_J = [0, 1, 2, 4, 8, 0, 1, 2, 3, 5, 12, 1, 6, 18, 0, 1, 7, 12, 1, 16, 1, 12, 1, 8, 18, 1, 16, 1, 3, 14, 18, 10, 16]
+    Table7_n = [
+        +0.801496989929495e-1,
+        -0.543862807146111,
+        +0.337455597421283,
+        +0.890555451157450e1,
+        +0.313840736431485e3,
+        +0.797367065977789,
+        -0.121616973556240e1,
+        +0.872803386937477e1,
+        -0.169769781757602e2,
+        -0.186552827328416e3,
+        +0.951159274344237e5,
+        -0.189168510120494e2,
+        -0.433407037194840e4,
+        +0.543212633012715e9,
+        +0.144793408386013,
+        +0.128024559637516e3,
+        -0.672309534071268e5,
+        +0.336972380095287e8,
+        -0.586634196762720e3,
+        -0.221403224769889e11,
+        +0.171606668708389e4,
+        -0.570817595806302e9,
+        -0.312109693178482e4,
+        -0.207841384633010e7,
+        +0.305605946157786e13,
+        +0.322157004314333e4,
+        +0.326810259797295e12,
+        -0.144104158934487e4,
+        +0.410694867802691e3,
+        +0.109077066873024e12,
+        -0.247964654258893e14,
+        +0.188801906865134e10,
+        -0.123651009018773e15,
+    ]
+
+    # Table 8. Coefficients and exponents of the backward equation p_2c(h,s) for subregion 2c in its dimensionless form, Eq. (5)
+    Table8_I = [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 4, 5, 5, 5, 5, 6, 6, 10, 12, 16]
+    Table8_J = [0, 1, 2, 3, 4, 8, 0, 2, 5, 8, 14, 2, 3, 7, 10, 18, 0, 5, 8, 16, 18, 18, 1, 4, 6, 14, 8, 18, 7, 7, 10]
+    Table8_n = [
+        +0.112225607199012,
+        -0.339005953606712e1,
+        -0.320503911730094e2,
+        -0.197597305104900e3,
+        -0.407693861553446e3,
+        +0.132943775222331e5,
+        +0.170846839774007e1,
+        +0.373694198142245e2,
+        +0.358144365815434e4,
+        +0.423014446424664e6,
+        -0.751071025760063e9,
+        +0.523446127607898e2,
+        -0.228351290812417e3,
+        -0.960652417056937e6,
+        -0.807059292526074e8,
+        +0.162698017225669e13,
+        +0.772465073604171,
+        +0.463929973837746e5,
+        -0.137317885134128e8,
+        +0.170470392630512e13,
+        -0.251104628187308e14,
+        +0.317748830835520e14,
+        +0.538685623675312e2,
+        -0.553089094625169e5,
+        -0.102861522421405e7,
+        +0.204249418756234e13,
+        +0.273918446626977e9,
+        -0.263963146312685e16,
+        -0.107890854108088e10,
+        -0.296492620980124e11,
+        -0.111754907323424e16,
+    ]
