+++ pyXSteam/tables/SR3_03.py
@@ -1,0 +1,870 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Content of the Tables from related documents
+
+Sources:
+
+* IAPWS SR3-03(2014) Revised Supplementary Release on Backward Equations for 
+the Functions T(p,h), v(p,h) and T(p,s), v(p,s) for Region 3 of the IAPWS Industrial
+Formulation 1997 for the Thermodynamic Properties of Water and Steam
+
+"""
+
+
+class SR3_03:
+    """class with table data from IAPWS SR3-03(2014)"""
+
+    s_c = 4.41202148223476  # kJ / kg K
+    h_dash = 1.670858218e3  # kJ / kg @ 623,15K
+    h_doubledash = 2.563592004e3  # kJ / kg @ 623,15K
+    s_dash = 3.778281340  # kJ / kg K @ 623,15K
+    s_doubledash = 5.210887825  # kJ / kg K @ 623,15K
+
+    # Table 2: Numerical values of the coefficients of the equation h_3ab(p) in its dimensionless
+    # form, Eq. (1), for defining the boundary between subregions 3a and 3b
+    Table2_n = [
+        +0.201464004206875e4,
+        +0.374696550136983e1,
+        -0.219921901054187e-1,
+        +0.875131686009950e-4,
+    ]
+
+    # Table 3: Coefficients and exponents of the backward equation T_3a(p,h) for subregion 3a in
+    # its dimensionless form, Eq. (2)
+    Table3_I = [
+        -12,
+        -12,
+        -12,
+        -12,
+        -12,
+        -12,
+        -12,
+        -12,
+        -10,
+        -10,
+        -10,
+        -8,
+        -8,
+        -8,
+        -8,
+        -5,
+        -3,
+        -2,
+        -2,
+        -2,
+        -1,
+        -1,
+        0,
+        0,
+        1,
+        3,
+        3,
+        4,
+        4,
+        10,
+        12,
+    ]
+    Table3_J = [
+        0,
+        1,
+        2,
+        6,
+        14,
+        16,
+        20,
+        22,
+        1,
+        5,
+        12,
+        0,
+        2,
+        4,
+        10,
+        2,
+        0,
+        1,
+        3,
+        4,
+        0,
+        2,
+        0,
+        1,
+        1,
+        0,
+        1,
+        0,
+        3,
+        4,
+        5,
+    ]
+    Table3_n = [
+        -0.133645667811215e-6,
+        +0.455912656802978e-5,
+        -0.146294640700979e-4,
+        +0.639341312970080e-2,
+        +0.372783927268847e3,
+        -0.718654377460447e4,
+        +0.573494752103400e6,
+        -0.267569329111439e7,
+        -0.334066283302614e-4,
+        -0.245479214069597e-1,
+        +0.478087847764996e2,
+        +0.764664131818904e-5,
+        +0.128350627676972e-2,
+        +0.171219081377331e-1,
+        -0.851007304583213e1,
+        -0.136513461629781e-1,
+        -0.384460997596657e-5,
+        +0.337423807911655e-2,
+        -0.551624873066791,
+        +0.729202277107470,
+        -0.992522757376041e-2,
+        -0.119308831407288,
+        +0.793929190615421,
+        +0.454270731799386,
+        +0.209998591259910,
+        -0.642109823904738e-2,
+        -0.235155868604540e-1,
+        +0.252233108341612e-2,
+        -0.764885133368119e-2,
+        +0.136176427574291e-1,
+        -0.133027883575669e-1,
+    ]
+
+    # Table 4: Coefficients and exponents of the backward equation T_3b(p,h) for subregion 3b in its dimensionless form, Eq. (3)
+    Table4_I = [
+        -12,
+        -12,
+        -10,
+        -10,
+        -10,
+        -10,
+        -10,
+        -8,
+        -8,
+        -8,
+        -8,
+        -8,
+        -6,
+        -6,
+        -6,
+        -4,
+        -4,
+        -3,
+        -2,
+        -2,
+        -1,
+        -1,
+        -1,
+        -1,
+        -1,
+        -1,
+        0,
+        0,
+        1,
+        3,
+        5,
+        6,
+        8,
+    ]
+    Table4_J = [
+        0,
+        1,
+        0,
+        1,
+        5,
+        10,
+        12,
+        0,
+        1,
+        2,
+        4,
+        10,
+        0,
+        1,
+        2,
+        0,
+        1,
+        5,
+        0,
+        4,
+        2,
+        4,
+        6,
+        10,
+        14,
+        16,
+        0,
+        2,
+        1,
+        1,
+        1,
+        1,
+        1,
+    ]
+    Table4_n = [
+        +0.323254573644920e-4,
+        -0.127575556587181e-3,
+        -0.475851877356068e-3,
+        +0.156183014181602e-2,
+        +0.105724860113781,
+        -0.858514221132534e2,
+        +0.724140095480911e3,
+        +0.296475810273257e-2,
+        -0.592721983365988e-2,
+        -0.126305422818666e-1,
+        -0.115716196364853,
+        +0.849000969739595e2,
+        -0.108602260086615e-1,
+        +0.154304475328851e-1,
+        +0.750455441524466e-1,
+        +0.252520973612982e-1,
+        -0.602507901232996e-1,
+        -0.307622221350501e1,
+        -0.574011959864879e-1,
+        +0.503471360939849e1,
+        -0.925081888584834,
+        +0.391733882917546e1,
+        -0.773146007130190e2,
+        +0.949308762098587e4,
+        -0.141043719679409e7,
+        +0.849166230819026e7,
+        +0.861095729446704,
+        +0.323346442811720,
+        +0.873281936020439,
+        -0.436653048526683,
+        +0.286596714529479,
+        -0.131778331276228,
+        +0.676682064330275e-2,
+    ]
+
+    # Table 6: Coefficients and exponents of the backward equation v_3a(p,h) for subregion 3a in its dimensionless form, Eq. (4)
+    Table6_I = [
+        -12,
+        -12,
+        -12,
+        -12,
+        -10,
+        -10,
+        -10,
+        -8,
+        -8,
+        -6,
+        -6,
+        -6,
+        -4,
+        -4,
+        -3,
+        -2,
+        -2,
+        -1,
+        -1,
+        -1,
+        -1,
+        0,
+        0,
+        1,
+        1,
+        1,
+        2,
+        2,
+        3,
+        4,
+        5,
+        8,
+    ]
+    Table6_J = [
+        6,
+        8,
+        12,
+        18,
+        4,
+        7,
+        10,
+        5,
+        12,
+        3,
+        4,
+        22,
+        2,
+        3,
+        7,
+        3,
+        16,
+        0,
+        1,
+        2,
+        3,
+        0,
+        1,
+        0,
+        1,
+        2,
+        0,
+        2,
+        0,
+        2,
+        2,
+        2,
+    ]
+    Table6_n = [
+        +0.529944062966028e-2,
+        -0.170099690234461,
+        +0.111323814312927e2,
+        -0.217898123145125e4,
+        -0.506061827980875e-3,
+        +0.556495239685324,
+        -0.943672726094016e1,
+        -0.297856807561527,
+        +0.939353943717186e2,
+        +0.192944939465981e-1,
+        +0.421740664704763,
+        -0.368914126282330e7,
+        -0.737566847600639e-2,
+        -0.354753242424366,
+        -0.199768169338727e1,
+        +0.115456297059049e1,
+        +0.568366875815960e4,
+        +0.808169540124668e-2,
+        +0.172416341519307,
+        +0.104270175292927e1,
+        -0.297691372792847,
+        +0.560394465163593,
+        +0.275234661176914,
+        -0.148347894866012,
+        -0.651142513478515e-1,
+        -0.292468715386302e1,
+        +0.664876096952665e-1,
+        +0.352335014263844e1,
+        -0.146340792313332e-1,
+        -0.224503486668184e1,
+        +0.110533464706142e1,
+        -0.408757344495612e-1,
+    ]
+
+    # Table 7: Coefficients and exponents of the backward equation v_3b(p,h) for subregion 3b in its dimensionless form, Eq. (5)
+    Table7_I = [
+        -12,
+        -12,
+        -8,
+        -8,
+        -8,
+        -8,
+        -8,
+        -8,
+        -6,
+        -6,
+        -6,
+        -6,
+        -6,
+        -6,
+        -4,
+        -4,
+        -4,
+        -3,
+        -3,
+        -2,
+        -2,
+        -1,
+        -1,
+        -1,
+        -1,
+        0,
+        1,
+        1,
+        2,
+        2,
+    ]
+    Table7_J = [
+        0,
+        1,
+        0,
+        1,
+        3,
+        6,
+        7,
+        8,
+        0,
+        1,
+        2,
+        5,
+        6,
+        10,
+        3,
+        6,
+        10,
+        0,
+        2,
+        1,
+        2,
+        0,
+        1,
+        4,
+        5,
+        0,
+        0,
+        1,
+        2,
+        6,
+    ]
+    Table7_n = [
+        -0.225196934336318e-8,
+        +0.140674363313486e-7,
+        +0.233784085280560e-5,
+        -0.331833715229001e-4,
+        +0.107956778514318e-2,
+        -0.271382067378863,
+        +0.107202262490333e1,
+        -0.853821329075382,
+        -0.215214194340526e-4,
+        +0.769656088222730e-3,
+        -0.431136580433864e-2,
+        +0.453342167309331,
+        -0.507749535873652,
+        -0.100475154528389e3,
+        -0.219201924648793,
+        -0.321087965668917e1,
+        +0.607567815637771e3,
+        +0.557686450685932e-3,
+        +0.187499040029550,
+        +0.905368030448107e-2,
+        +0.285417173048685,
+        +0.329924030996098e-1,
+        +0.239897419685483,
+        +0.482754995951394e1,
+        -0.118035753702231e2,
+        +0.169490044091791,
+        -0.179967222507787e-1,
+        +0.371810116332674e-1,
+        -0.536288335065096e-1,
+        +0.160697101092520e1,
+    ]
+
+    # Table 10: Coefficients and exponents of the backward equation T_3a(p,s) for subregion 3a in its dimensionless form, Eq. (6)
+    Table10_I = [
+        -12,
+        -12,
+        -10,
+        -10,
+        -10,
+        -10,
+        -8,
+        -8,
+        -8,
+        -8,
+        -6,
+        -6,
+        -6,
+        -5,
+        -5,
+        -5,
+        -4,
+        -4,
+        -4,
+        -2,
+        -2,
+        -1,
+        -1,
+        0,
+        0,
+        0,
+        1,
+        2,
+        2,
+        3,
+        8,
+        8,
+        10,
+    ]
+    Table10_J = [
+        28,
+        32,
+        4,
+        10,
+        12,
+        14,
+        5,
+        7,
+        8,
+        28,
+        2,
+        6,
+        32,
+        0,
+        14,
+        32,
+        6,
+        10,
+        36,
+        1,
+        4,
+        1,
+        6,
+        0,
+        1,
+        4,
+        0,
+        0,
+        3,
+        2,
+        0,
+        1,
+        2,
+    ]
+    Table10_n = [
+        +0.150042008263875e10,
+        -0.159397258480424e12,
+        +0.502181140217975e-3,
+        -0.672057767855466e2,
+        +0.145058545404456e4,
+        -0.823889534888890e4,
+        -0.154852214233853,
+        +0.112305046746695e2,
+        -0.297000213482822e2,
+        +0.438565132635495e11,
+        +0.137837838635464e-2,
+        -0.297478527157462e1,
+        +0.971777947349413e13,
+        -0.571527767052398e-4,
+        +0.288307949778420e5,
+        -0.744428289262703e14,
+        +0.128017324848921e2,
+        -0.368275545889071e3,
+        +0.664768904779177e16,
+        +0.449359251958880e-1,
+        -0.422897836099655e1,
+        -0.240614376434179,
+        -0.474341365254924e1,
+        +0.724093999126110,
+        +0.923874349695897,
+        +0.399043655281015e1,
+        +0.384066651868009e-1,
+        -0.359344365571848e-2,
+        -0.735196448821653,
+        +0.188367048396131,
+        +0.141064266818704e-3,
+        -0.257418501496337e-2,
+        +0.123220024851555e-2,
+    ]
+
+    # Table 11: Coefficients and exponents of the backward equation T_3b(p,s) for subregion 3b in its dimensionless form, Eq. (7)
+    Table11_I = [
+        -12,
+        -12,
+        -12,
+        -12,
+        -8,
+        -8,
+        -8,
+        -6,
+        -6,
+        -6,
+        -5,
+        -5,
+        -5,
+        -5,
+        -5,
+        -4,
+        -3,
+        -3,
+        -2,
+        0,
+        2,
+        3,
+        4,
+        5,
+        6,
+        8,
+        12,
+        14,
+    ]
+    Table11_J = [
+        1,
+        3,
+        4,
+        7,
+        0,
+        1,
+        3,
+        0,
+        2,
+        4,
+        0,
+        1,
+        2,
+        4,
+        6,
+        12,
+        1,
+        6,
+        2,
+        0,
+        1,
+        1,
+        0,
+        24,
+        0,
+        3,
+        1,
+        2,
+    ]
+    Table11_n = [
+        +0.527111701601660,
+        -0.401317830052742e2,
+        +0.153020073134484e3,
+        -0.224799398218827e4,
+        -0.193993484669048,
+        -0.140467557893768e1,
+        +0.426799878114024e2,
+        +0.752810643416743,
+        +0.226657238616417e2,
+        -0.622873556909932e3,
+        -0.660823667935396,
+        +0.841267087271658,
+        -0.253717501764397e2,
+        +0.485708963532948e3,
+        +0.880531517490555e3,
+        +0.265015592794626e7,
+        -0.359287150025783,
+        -0.656991567673753e3,
+        +0.241768149185367e1,
+        +0.856873461222588,
+        +0.655143675313458,
+        -0.213535213206406,
+        +0.562974957606348e-2,
+        -0.316955725450471e15,
+        -0.699997000152457e-3,
+        +0.119845803210767e-1,
+        +0.193848122022095e-4,
+        -0.215095749182309e-4,
+    ]
+
+    # Table 13: Coefficients and exponents of the backward equation v_3a(p,s) for subregion 3a in its dimensionless form, Eq. (8).
+    Table13_I = [
+        -12,
+        -12,
+        -12,
+        -10,
+        -10,
+        -10,
+        -10,
+        -8,
+        -8,
+        -8,
+        -8,
+        -6,
+        -5,
+        -4,
+        -3,
+        -3,
+        -2,
+        -2,
+        -1,
+        -1,
+        0,
+        0,
+        0,
+        1,
+        2,
+        4,
+        5,
+        6,
+    ]
+    Table13_J = [
+        10,
+        12,
+        14,
+        4,
+        8,
+        10,
+        20,
+        5,
+        6,
+        14,
+        16,
+        28,
+        1,
+        5,
+        2,
+        4,
+        3,
+        8,
+        1,
+        2,
+        0,
+        1,
+        3,
+        0,
+        0,
+        2,
+        2,
+        0,
+    ]
+    Table13_n = [
+        +0.795544074093975e2,
+        -0.238261242984590e4,
+        +0.176813100617787e5,
+        -0.110524727080379e-2,
+        -0.153213833655326e2,
+        +0.297544599376982e3,
+        -0.350315206871242e8,
+        +0.277513761062119,
+        -0.523964271036888,
+        -0.148011182995403e6,
+        +0.160014899374266e7,
+        +0.170802322663427e13,
+        +0.246866996006494e-3,
+        +0.165326084797980e1,
+        -0.118008384666987,
+        +0.253798642355900e1,
+        +0.965127704669424,
+        -0.282172420532826e2,
+        +0.203224612353823,
+        +0.110648186063513e1,
+        +0.526127948451280,
+        +0.277000018736321,
+        +0.108153340501132e1,
+        -0.744127885357893e-1,
+        +0.164094443541384e-1,
+        -0.680468275301065e-1,
+        +0.257988576101640e-1,
+        -0.145749861944416e-3,
+    ]
+
+    # Table 14: Coefficients and exponents of the backward equation v_3b(p,s) for subregion 3b in its dimensionless form, Eq. (9)
+    Table14_I = [
+        -12,
+        -12,
+        -12,
+        -12,
+        -12,
+        -12,
+        -10,
+        -10,
+        -10,
+        -10,
+        -8,
+        -5,
+        -5,
+        -5,
+        -4,
+        -4,
+        -4,
+        -4,
+        -3,
+        -2,
+        -2,
+        -2,
+        -2,
+        -2,
+        -2,
+        0,
+        0,
+        0,
+        1,
+        1,
+        2,
+    ]
+    Table14_J = [
+        0,
+        1,
+        2,
+        3,
+        5,
+        6,
+        0,
+        1,
+        2,
+        4,
+        0,
+        1,
+        2,
+        3,
+        0,
+        1,
+        2,
+        3,
+        1,
+        0,
+        1,
+        2,
+        3,
+        4,
+        12,
+        0,
+        1,
+        2,
+        0,
+        2,
+        2,
+    ]
+    Table14_n = [
+        +0.591599780322238e-4,
+        -0.185465997137856e-2,
+        +0.104190510480013e-1,
+        +0.598647302038590e-2,
+        -0.771391189901699,
+        +0.172549765557036e1,
+        -0.467076079846526e-3,
+        +0.134533823384439e-1,
+        -0.808094336805495e-1,
+        +0.508139374365767,
+        +0.128584643361683e-2,
+        -0.163899353915435e1,
+        +0.586938199318063e1,
+        -0.292466667918613e1,
+        -0.614076301499537e-2,
+        +0.576199014049172e1,
+        -0.121613320606788e2,
+        +0.167637540957944e1,
+        -0.744135838773463e1,
+        +0.378168091437659e-1,
+        +0.401432203027688e1,
+        +0.160279837479185e2,
+        +0.317848779347728e1,
+        -0.358362310304853e1,
+        -0.115995260446827e7,
+        +0.199256573577909,
+        -0.122270624794624,
+        -0.191449143716586e2,
+        -0.150448002905284e-1,
+        +0.146407900162154e2,
+        -0.327477787188230e1,
+    ]
+
+    # Table 17: Coefficients and exponents of the boundary equation p_3sat(h) in its dimensionless form, Eq. (10)
+    Table17_I = [0, 1, 1, 1, 1, 5, 7, 8, 14, 20, 22, 24, 28, 36]
+    Table17_J = [0, 1, 3, 4, 36, 3, 0, 24, 16, 16, 3, 18, 8, 24]
+    Table17_n = [
+        +0.600073641753024,
+        -0.936203654849857e1,
+        +0.246590798594147e2,
+        -0.107014222858224e3,
+        -0.915821315805768e14,
+        -0.862332011700662e4,
+        -0.235837344740032e2,
+        +0.252304969384128e18,
+        -0.389718771997719e19,
+        -0.333775713645296e23,
+        +0.356499469636328e11,
+        -0.148547544720641e27,
+        +0.330611514838798e19,
+        +0.813641294467829e38,
+    ]
+
+    # Table 19: Coefficients and exponents of the boundary equation p_3sat(s) in its dimensionless form, Eq. (11)
+    Table19_I = [0, 1, 1, 4, 12, 12, 16, 24, 28, 32]
+    Table19_J = [0, 1, 32, 7, 4, 14, 36, 10, 0, 18]
+    Table19_n = [
+        +0.639767553612785,
+        -0.129727445396014e2,
+        -0.224595125848403e16,
+        +0.177466741801846e7,
+        +0.717079349571538e10,
+        -0.378829107169011e18,
+        -0.955586736431328e35,
+        +0.187269814676188e24,
+        +0.119254746466473e12,
+        +0.110649277244882e37,
+    ]
