+++ pyXSteam/tables/SR4_04.py
@@ -1,0 +1,421 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Content of the Tables from related documents
+
+Sources:
+
+* IAPWS SR4-04(2014) Revised Supplementary Release on Backward Equations p(h,s) 
+for Region 3, Equations as a Function of h and s for the Region Boundaries, and 
+an Equation Tsat(h,s) for Region 4 of the IAPWS Industrial Formulation 1997 for 
+the Thermodynamic Properties of Water and Steam 
+
+"""
+
+from .SR3_03 import SR3_03
+
+
+class SR4_04:
+    """class with table data from IAPWS SR4-04(2014)"""
+
+    s_dash_low = -1.545495919e-4  # kJ / kg K @ 273.15K
+    s_dash_heigh = 3.778281340  # kJ / kg K @ 623.15K
+    s_c = SR3_03.s_c  # 4.41202148223476  # kJ / kg K @ 623.15K
+    s_doubledash = 9.155759395  # kJ / kg K @ 273.15K
+    s_heigh_mpa = 3.397782955  # kJ / kg K @ 100MPA, 673.15K
+    s_min_B23 = 5.048096828  # kJ / kg K
+    s_max_B23 = 5.260578707  # kJ / kg K
+
+    h_min_B23 = SR3_03.h_doubledash  # kJ / kg
+    h_max_B23 = 2.812942061e2  # kJ / kg
+
+    # Table 3: Coefficients and exponents of the backward equation p_3a (h,s) for subregion 3a in its dimensionless form, Eq. (1)
+    Table3_I = [0, 0, 0, 1, 1, 1, 1, 1, 2, 2, 3, 3, 3, 4, 4, 4, 4, 5, 6, 7, 8, 10, 10, 14, 18, 20, 22, 22, 24, 28, 28, 32, 32]
+    Table3_J = [0, 1, 5, 0, 3, 4, 8, 14, 6, 16, 0, 2, 3, 0, 1, 4, 5, 28, 28, 24, 1, 32, 36, 22, 28, 36, 16, 28, 36, 16, 36, 10, 28]
+    Table3_n = [
+        +0.770889828326934e1,
+        -0.260835009128688e2,
+        +0.267416218930389e3,
+        +0.172221089496844e2,
+        -0.293542332145970e3,
+        +0.614135601882478e3,
+        -0.610562757725674e5,
+        -0.651272251118219e8,
+        +0.735919313521937e5,
+        -0.116646505914191e11,
+        +0.355267086434461e2,
+        -0.596144543825955e3,
+        -0.475842430145708e3,
+        +0.696781965359503e2,
+        +0.335674250377312e3,
+        +0.250526809130882e5,
+        +0.146997380630766e6,
+        +0.538069315091534e20,
+        +0.143619827291346e22,
+        +0.364985866165994e20,
+        -0.254741561156775e4,
+        +0.240120197096563e28,
+        -0.393847464679496e30,
+        +0.147073407024852e25,
+        -0.426391250432059e32,
+        +0.194509340621077e39,
+        +0.666212132114896e24,
+        +0.706777016552858e34,
+        +0.175563621975576e42,
+        +0.108408607429124e29,
+        +0.730872705175151e44,
+        +0.159145847398870e25,
+        +0.377121605943324e41,
+    ]
+
+    # Table 4: Coefficients and exponents of the backward equation p_3b (h,s) for subregion 3b in its dimensionless form, Eq. (2)
+    Table4_I = [
+        -12,
+        -12,
+        -12,
+        -12,
+        -12,
+        -10,
+        -10,
+        -10,
+        -10,
+        -8,
+        -8,
+        -6,
+        -6,
+        -6,
+        -6,
+        -5,
+        -4,
+        -4,
+        -4,
+        -3,
+        -3,
+        -3,
+        -3,
+        -2,
+        -2,
+        -1,
+        0,
+        2,
+        2,
+        5,
+        6,
+        8,
+        10,
+        14,
+        14,
+    ]
+    Table4_J = [2, 10, 12, 14, 20, 2, 10, 14, 18, 2, 8, 2, 6, 7, 8, 10, 4, 5, 8, 1, 3, 5, 6, 0, 1, 0, 3, 0, 1, 0, 1, 1, 1, 3, 7]
+    Table4_n = [
+        +0.125244360717979e-12,
+        -0.126599322553713e-1,
+        +0.506878030140626e1,
+        +0.317847171154202e2,
+        -0.391041161399932e6,
+        -0.975733406392044e-10,
+        -0.186312419488279e2,
+        +0.510973543414101e3,
+        +0.373847005822362e6,
+        +0.299804024666572e-7,
+        +0.200544393820342e2,
+        -0.498030487662829e-5,
+        -0.102301806360030e2,
+        +0.552819126990325e2,
+        -0.206211367510878e3,
+        -0.794012232324823e4,
+        +0.782248472028153e1,
+        -0.586544326902468e2,
+        +0.355073647696481e4,
+        -0.115303107290162e-3,
+        -0.175092403171802e1,
+        +0.257981687748160e3,
+        -0.727048374179467e3,
+        +0.121644822609198e-3,
+        +0.393137871762692e-1,
+        +0.704181005909296e-2,
+        -0.829108200698110e2,
+        -0.265178818131250,
+        +0.137531682453991e2,
+        -0.522394090753046e2,
+        +0.240556298941048e4,
+        -0.227361631268929e5,
+        +0.890746343932567e5,
+        -0.239234565822486e8,
+        +0.568795808129714e10,
+    ]
+
+    # Table 9: Coefficients and exponents of the boundary equation h'_1 (s) in its dimensionless form, Eq. (3)
+    Table9_I = [0, 0, 1, 1, 2, 2, 3, 3, 4, 4, 4, 5, 5, 7, 8, 12, 12, 14, 14, 16, 20, 20, 22, 24, 28, 32, 32]
+    Table9_J = [14, 36, 3, 16, 0, 5, 4, 36, 4, 16, 24, 18, 24, 1, 4, 2, 4, 1, 22, 10, 12, 28, 8, 3, 0, 6, 8]
+    Table9_n = [
+        +0.332171191705237,
+        +0.611217706323496e-3,
+        -0.882092478906822e1,
+        -0.455628192543250,
+        -0.263483840850452e-4,
+        -0.223949661148062e2,
+        -0.428398660164013e1,
+        -0.616679338856916,
+        -0.146823031104040e2,
+        +0.284523138727299e3,
+        -0.113398503195444e3,
+        +0.115671380760859e4,
+        +0.395551267359325e3,
+        -0.154891257229285e1,
+        +0.194486637751291e2,
+        -0.357915139457043e1,
+        -0.335369414148819e1,
+        -0.664426796332460,
+        +0.323321885383934e5,
+        +0.331766744667084e4,
+        -0.223501257931087e5,
+        +0.573953875852936e7,
+        +0.173226193407919e3,
+        -0.363968822121321e-1,
+        +0.834596332878346e-6,
+        +0.503611916682674e1,
+        +0.655444787064505e2,
+    ]
+
+    # Table 10: Coefficients and exponents of the boundary equation h'_3a(s) in its dimensionless form, Eq. (4)
+    Table10_I = [0, 0, 0, 0, 2, 3, 4, 4, 5, 5, 6, 7, 7, 7, 10, 10, 10, 32, 32]
+    Table10_J = [1, 4, 10, 16, 1, 36, 3, 16, 20, 36, 4, 2, 28, 32, 14, 32, 36, 0, 6]
+    Table10_n = [
+        +0.822673364673336,
+        +0.181977213534479,
+        -0.112000260313624e-1,
+        -0.746778287048033e-3,
+        -0.179046263257381,
+        +0.424220110836657e-1,
+        -0.341355823438768,
+        -0.209881740853565e1,
+        -0.822477343323596e1,
+        -0.499684082076008e1,
+        +0.191413958471069,
+        +0.581062241093136e-1,
+        -0.165505498701029e4,
+        +0.158870443421201e4,
+        -0.850623535172818e2,
+        -0.317714386511207e5,
+        -0.945890406632871e5,
+        -0.13927384708869e-5,
+        +0.63105253224098,
+    ]
+
+    # Table 16: Coefficients and exponents of the boundary equation h''_2ab(s) in its dimensionless form, Eq. (5)
+    Table16_I = [1, 1, 2, 2, 4, 4, 7, 8, 8, 10, 12, 12, 18, 20, 24, 28, 28, 28, 28, 28, 32, 32, 32, 32, 32, 36, 36, 36, 36, 36]
+    Table16_J = [8, 24, 4, 32, 1, 2, 7, 5, 12, 1, 0, 7, 10, 12, 32, 8, 12, 20, 22, 24, 2, 7, 12, 14, 24, 10, 12, 20, 22, 28]
+    Table16_n = [
+        -0.524581170928788e3,
+        -0.926947218142218e7,
+        -0.237385107491666e3,
+        +0.210770155812776e11,
+        -0.239494562010986e2,
+        +0.221802480294197e3,
+        -0.510472533393438e7,
+        +0.124981396109147e7,
+        +0.200008436996201e10,
+        -0.815158509791035e3,
+        -0.157612685637523e3,
+        -0.114200422332791e11,
+        +0.662364680776872e16,
+        -0.227622818296144e19,
+        -0.171048081348406e32,
+        +0.660788766938091e16,
+        +0.166320055886021e23,
+        -0.218003784381501e30,
+        -0.787276140295618e30,
+        +0.151062329700346e32,
+        +0.795732170300541e7,
+        +0.131957647355347e16,
+        -0.325097068299140e24,
+        -0.418600611419248e26,
+        +0.297478906557467e35,
+        -0.953588761745473e20,
+        +0.166957699620939e25,
+        -0.175407764869978e33,
+        +0.347581490626396e35,
+        -0.710971318427851e39,
+    ]
+
+    # Table 17: Coefficients and exponents of the boundary equation h''_2c3b(s) in its dimensionless form, Eq. (6)
+    Table17_I = [0, 0, 0, 1, 1, 5, 6, 7, 8, 8, 12, 16, 22, 22, 24, 36]
+    Table17_J = [0, 3, 4, 0, 12, 36, 12, 16, 2, 20, 32, 36, 2, 32, 7, 20]
+    Table17_n = [
+        +0.104351280732769e1,
+        -0.227807912708513e1,
+        +0.180535256723202e1,
+        +0.420440834792042,
+        -0.105721244834660e6,
+        +0.436911607493884e25,
+        -0.328032702839753e12,
+        -0.678686760804270e16,
+        +0.743957464645363e4,
+        -0.356896445355761e20,
+        +0.167590585186801e32,
+        -0.355028625419105e38,
+        +0.396611982166538e12,
+        -0.414716268484468e41,
+        +0.359080103867382e19,
+        -0.116994334851995e41,
+    ]
+
+    # Table 23. Coefficients and exponents of the equation h_B13(s) in its dimensionless form, Eq. (7)
+    Table23_I = [0, 1, 1, 3, 5, 6]
+    Table23_J = [0, -2, 2, -12, -4, -3]
+    Table23_n = [
+        +0.913965547600543,
+        -0.430944856041991e-4,
+        +0.603235694765419e2,
+        +0.117518273082168e-17,
+        +0.220000904781292,
+        -0.690815545851641e2,
+    ]
+
+    # Table 25: Coefficients and exponents of the boundary equation T_B23(h,s) in its dimensionless form, Eq. (8)
+    Table25_I = [-12, -10, -8, -4, -3, -2, -2, -2, -2, 0, 1, 1, 1, 3, 3, 5, 6, 6, 8, 8, 8, 12, 12, 14, 14]
+    Table25_J = [10, 8, 3, 4, 3, -6, 2, 3, 4, 0, -3, -2, 10, -2, -1, -5, -6, -3, -8, -2, -1, -12, -1, -12, 1]
+    Table25_n = [
+        +0.629096260829810e-3,
+        -0.823453502583165e-3,
+        +0.515446951519474e-7,
+        -0.117565945784945e1,
+        +0.348519684726192e1,
+        -0.507837382408313e-11,
+        -0.284637670005479e1,
+        -0.236092263939673e1,
+        +0.601492324973779e1,
+        +0.148039650824546e1,
+        +0.360075182221907e-3,
+        -0.126700045009952e-1,
+        -0.122184332521413e7,
+        +0.149276502463272,
+        +0.698733471798484,
+        -0.252207040114321e-1,
+        +0.147151930985213e-1,
+        -0.108618917681849e1,
+        -0.936875039816322e-3,
+        +0.819877897570217e2,
+        -0.182041861521835e3,
+        +0.261907376402688e-5,
+        -0.291626417025961e5,
+        +0.140660774926165e-4,
+        +0.783237062349385e7,
+    ]
+
+    # Table 28: Coefficients and exponents of the equation T_sat(h,s) in its dimensionless form, Eq. (9)
+    Table28_I = [
+        0,
+        0,
+        0,
+        1,
+        1,
+        1,
+        1,
+        2,
+        2,
+        2,
+        3,
+        3,
+        3,
+        3,
+        4,
+        4,
+        5,
+        5,
+        5,
+        5,
+        6,
+        6,
+        6,
+        8,
+        10,
+        10,
+        12,
+        14,
+        14,
+        16,
+        16,
+        18,
+        18,
+        18,
+        20,
+        28,
+    ]
+    Table28_J = [
+        0,
+        3,
+        12,
+        0,
+        1,
+        2,
+        5,
+        0,
+        5,
+        8,
+        0,
+        2,
+        3,
+        4,
+        0,
+        1,
+        1,
+        2,
+        4,
+        16,
+        6,
+        8,
+        22,
+        1,
+        20,
+        36,
+        24,
+        1,
+        28,
+        12,
+        32,
+        14,
+        22,
+        36,
+        24,
+        36,
+    ]
+    Table28_n = [
+        +0.179882673606601,
+        -0.267507455199603,
+        +0.116276722612600e1,
+        +0.147545428713616,
+        -0.512871635973248,
+        +0.421333567697984,
+        +0.563749522189870,
+        +0.429274443819153,
+        -0.335704552142140e1,
+        +0.108890916499278e2,
+        -0.248483390456012,
+        +0.304153221906390,
+        -0.494819763939905,
+        +0.107551674933261e1,
+        +0.733888415457688e-1,
+        +0.140170545411085e-1,
+        -0.106110975998808,
+        +0.168324361811875e-1,
+        +0.125028363714877e1,
+        +0.101316840309509e4,
+        -0.151791558000712e1,
+        +0.524277865990866e2,
+        +0.230495545563912e5,
+        +0.249459806365456e-1,
+        +0.210796467412137e7,
+        +0.366836848613065e9,
+        -0.144814105365163e9,
+        -0.179276373003590e-2,
+        +0.489955602100459e10,
+        +0.471262212070518e3,
+        -0.829294390198652e11,
+        -0.171545662263191e4,
+        +0.355777682973575e7,
+        +0.586062760258436e12,
+        -0.129887635078195e8,
+        +0.317247449371057e11,
+    ]
