+++ pyXSteam/tables/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""module with all collected data from coefficient or other tables of related documents"""
+
+from .R7_97 import R7_97
+from .SR1_86 import SR1_86
+from .SR2_01 import SR2_01
+from .SR3_03 import SR3_03
+from .SR4_04 import SR4_04
